refactor(server): split match.ts — extract 5 details/list handlers (S27.8.19)
Fourth slice on `apps/server/src/routes/match.ts`. Extract the 5
"details / list" read-only handlers into a dedicated module
`routes/match-details-handlers.ts` (464 lines).

Endpoints covered:
- `GET /details` (token-based) — handleGetMatchDetailsByToken
- `GET /:id/details` — handleGetMatchDetails
- `GET /:id/teams` — handleGetMatchTeams
- `GET /my-matches` — handleListMyMatches
- `GET /:id/summary` — handleGetMatchSummary

All thematically read-only (team/coach names, ELO, lists, summary
metadata) and isolated from the lifecycle handlers that remain in
`match.ts`.

Re-export from `match.ts` to preserve the public API consumed by
integration tests.

Leaf dependencies only (`prisma`, `jwt`, `MATCH_SECRET`, `serverLog`,
`sendError/sendSuccess`). No cycle.

`match.ts` shrinks 1176 -> 809 lines (-367 net lines, cumulative
-957 vs origin 1766 since S27.8.16 = -54%).

- 1787/1787 server tests pass (no regression)
- typecheck clean
