+++ apps/server/src/routes/match.ts
@@ -278,397 +278,30 @@ router.post("/:id/move", authUser, validate(moveSchema), handleSubmitMove);
 
 export default router;
 
-// Détails du match: noms des équipes et coachs (via token de match)
-// (S25.5j — ApiResponse<T>)
-export async function handleGetMatchDetailsByToken(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const token = (req.headers["x-match-token"] as string) || "";
-    if (!token) {
-      sendError(res, "x-match-token requis", 401);
-      return;
-    }
-    let payload: any;
-    try {
-      payload = jwt.verify(token, MATCH_SECRET) as any;
-    } catch {
-      sendError(res, "x-match-token invalide", 401);
-      return;
-    }
-    const matchId = payload?.matchId as string | undefined;
-    if (!matchId) {
-      sendError(res, "matchId manquant dans le token", 400);
-      return;
-    }
+// S27.8.19 — Handlers de details (5 endpoints lecture seule de
+// metadonnees) extraits dans `routes/match-details-handlers.ts`.
+// Re-export pour preserver l'API publique consommee par les tests
+// d'integration.
+export {
+  handleGetMatchDetailsByToken,
+  handleGetMatchDetails,
+  handleGetMatchTeams,
+  handleListMyMatches,
+  handleGetMatchSummary,
+} from './match-details-handlers';
+import {
+  handleGetMatchDetailsByToken as handleGetMatchDetailsByTokenImpl,
+  handleGetMatchDetails as handleGetMatchDetailsImpl,
+  handleGetMatchTeams as handleGetMatchTeamsImpl,
+  handleListMyMatches as handleListMyMatchesImpl,
+  handleGetMatchSummary as handleGetMatchSummaryImpl,
+} from './match-details-handlers';
 
-    const [match, selections] = await Promise.all([
-      prisma.match.findUnique({
-        where: { id: matchId },
-        select: { id: true, creatorId: true },
-      }),
-      prisma.teamSelection.findMany({
-        where: { matchId },
-        orderBy: { createdAt: "asc" },
-        include: {
-          user: { select: { id: true, name: true, email: true } },
-          teamRef: { select: { name: true, roster: true } },
-        },
-      }),
-    ]);
-
-    const tokenUserId = (payload as any)?.userId as string | undefined;
-    let local = selections.find((s: any) => s.userId === tokenUserId) || null;
-    let visitor = selections.find((s: any) => s.userId !== tokenUserId) || null;
-    if (!local || !visitor) {
-      if (match?.creatorId) {
-        local =
-          selections.find((s: any) => s.userId === match.creatorId) || selections[0];
-        visitor =
-          selections.find((s: any) => s.userId !== match.creatorId) ||
-          selections[1] ||
-          null;
-      } else {
-        local = selections[0] || null;
-        visitor = selections.length > 1 ? selections[1] : null;
-      }
-    }
-
-    const teamName = (sel: any): string => {
-      if (!sel) return "";
-      return sel.teamRef?.name || sel.teamRef?.roster || sel.team || "";
-    };
-    const coachName = (sel: any): string => {
-      if (!sel) return "";
-      return sel.user?.name || sel.user?.email || "";
-    };
-
-    sendSuccess(res, {
-      matchId,
-      local: {
-        teamName: teamName(local),
-        coachName: coachName(local),
-        userId: local?.userId || null,
-      },
-      visitor: {
-        teamName: teamName(visitor),
-        coachName: coachName(visitor),
-        userId: visitor?.userId || null,
-      },
-    });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/details", handleGetMatchDetailsByToken);
-
-// Détails du match (auth) par id (S25.5j — ApiResponse<T>)
-export async function handleGetMatchDetails(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const matchId = req.params.id;
-    const [match, selections] = await Promise.all([
-      prisma.match.findUnique({
-        where: { id: matchId },
-        select: { id: true, creatorId: true },
-      }),
-      prisma.teamSelection.findMany({
-        where: { matchId },
-        orderBy: { createdAt: "asc" },
-        include: {
-          user: { select: { id: true, name: true, email: true, eloRating: true } },
-          teamRef: { select: { name: true, roster: true } },
-        },
-      }),
-    ]);
-    if (!match) {
-      sendError(res, "Partie introuvable", 404);
-      return;
-    }
-    const authenticatedUserId = req.user!.id;
-    let local =
-      selections.find((s: any) => s.userId === authenticatedUserId) || null;
-    let visitor =
-      selections.find((s: any) => s.userId !== authenticatedUserId) || null;
-    if (!local || !visitor) {
-      if (match.creatorId) {
-        local =
-          selections.find((s: any) => s.userId === match.creatorId) ||
-          selections[0] ||
-          null;
-        visitor =
-          selections.find((s: any) => s.userId !== match.creatorId) ||
-          selections[1] ||
-          null;
-      } else {
-        local = selections[0] || null;
-        visitor = selections.length > 1 ? selections[1] : null;
-      }
-    }
-    const teamName = (sel: any) =>
-      sel?.teamRef?.name || sel?.teamRef?.roster || sel?.team || "";
-    const coachName = (sel: any) => sel?.user?.name || sel?.user?.email || "";
-    const eloRating = (sel: any) => sel?.user?.eloRating ?? 1000;
-    sendSuccess(res, {
-      matchId,
-      local: { teamName: teamName(local), coachName: coachName(local), eloRating: eloRating(local) },
-      visitor: { teamName: teamName(visitor), coachName: coachName(visitor), eloRating: eloRating(visitor) },
-    });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/:id/details", authUser, handleGetMatchDetails);
-
-// Nouvel endpoint pour les équipes et joueurs (auth) par id - pour fallback prematch
-// IMPORTANT: retourne une vue absolue A/B, indépendante de l'utilisateur connecté
-// (S25.5j — ApiResponse<T>)
-export async function handleGetMatchTeams(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const matchId = req.params.id;
-    const match = await prisma.match.findUnique({
-      where: { id: matchId },
-      select: { id: true },
-    });
-    if (!match) {
-      sendError(res, "Partie introuvable", 404);
-      return;
-    }
-
-    const selections = await prisma.teamSelection.findMany({
-      where: { matchId },
-      orderBy: { createdAt: "asc" },
-      include: { teamRef: { include: { players: true } } },
-    });
-
-    const s1 = selections[0] || null;
-    const s2 = selections[1] || null;
-
-    const getTeamData = (sel: any) => {
-      const teamName = sel?.teamRef?.name || sel?.team || "Équipe Inconnue";
-      const players = (sel?.teamRef?.players || []).map((p: any) => ({
-        id: p.id,
-        name: p.name,
-        position: p.position,
-        number: p.number,
-        ma: p.ma,
-        st: p.st,
-        ag: p.ag,
-        pa: p.pa,
-        av: p.av,
-        skills: p.skills || "",
-      }));
-      return { teamName, players };
-    };
-
-    const teamA = getTeamData(s1);
-    const teamB = getTeamData(s2);
-
-    sendSuccess(res, { teamA, teamB });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/:id/teams", authUser, handleGetMatchTeams);
-
-// Liste des matchs de l'utilisateur connecté
-// Liste des matchs du joueur connecte (S25.5h — ApiResponse<T>)
-export async function handleListMyMatches(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const userId = req.user!.id;
-
-    const matches = await prisma.match.findMany({
-      where: {
-        players: { some: { id: userId } },
-      },
-      orderBy: { createdAt: "desc" },
-      take: 50,
-      include: {
-        teamSelections: {
-          include: {
-            user: { select: { id: true, coachName: true, eloRating: true } },
-            teamRef: { select: { id: true, name: true, roster: true } },
-          },
-        },
-        turns: {
-          orderBy: { number: "desc" },
-          take: 1,
-          select: { payload: true },
-        },
-      },
-    });
-
-    const result = matches.map((m: any) => {
-      // Extraire le score depuis le dernier turn si disponible
-      const lastTurn = m.turns[0];
-      const gameState = (lastTurn?.payload as any)?.gameState;
-      const score = gameState?.score || { teamA: 0, teamB: 0 };
-      const half = gameState?.half || 0;
-      const turn = gameState?.turn || 0;
-
-      // Déterminer le côté de l'utilisateur
-      const selections = m.teamSelections.sort(
-        (a: any, b: any) => a.createdAt.getTime() - b.createdAt.getTime(),
-      );
-      const mySelection = selections.find((s: any) => s.userId === userId);
-      const opponentSelection = selections.find((s: any) => s.userId !== userId);
-      const isMyTurn = m.currentTurnUserId === userId;
-
-      // Align score with user's side: first selection = teamA, second = teamB
-      const isTeamA = mySelection === selections[0];
-      const myScore = isTeamA ? score.teamA : score.teamB;
-      const opponentScore = isTeamA ? score.teamB : score.teamA;
-
-      return {
-        id: m.id,
-        status: m.status,
-        createdAt: m.createdAt,
-        lastMoveAt: m.lastMoveAt,
-        isMyTurn,
-        score,
-        myScore,
-        opponentScore,
-        half,
-        turn,
-        myTeam: mySelection
-          ? {
-              coachName: mySelection.user.coachName,
-              teamName: mySelection.teamRef?.name || mySelection.team,
-              rosterName: mySelection.teamRef?.roster,
-              eloRating: mySelection.user.eloRating,
-            }
-          : null,
-        opponent: opponentSelection
-          ? {
-              coachName: opponentSelection.user.coachName,
-              teamName: opponentSelection.teamRef?.name || opponentSelection.team,
-              rosterName: opponentSelection.teamRef?.roster,
-              eloRating: opponentSelection.user.eloRating,
-            }
-          : null,
-      };
-    });
-
-    sendSuccess(res, { matches: result });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/my-matches", authUser, handleListMyMatches);
-
-// Résumé d'un match: équipes, coachs, score (approx), tour/mi-temps
-// (S25.5k — ApiResponse<T>)
-export async function handleGetMatchSummary(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  try {
-    const matchId = req.params.id;
-    const match = await prisma.match.findUnique({
-      where: { id: matchId },
-      select: {
-        id: true,
-        status: true,
-        seed: true,
-        creatorId: true,
-        createdAt: true,
-      },
-    });
-    if (!match) {
-      sendError(res, "Partie introuvable", 404);
-      return;
-    }
-
-    const [selections, acceptTurns] = await Promise.all([
-      prisma.teamSelection.findMany({
-        where: { matchId },
-        include: {
-          user: { select: { id: true, name: true, email: true, eloRating: true } },
-          teamRef: { select: { id: true, name: true, roster: true } },
-        },
-        orderBy: { createdAt: "asc" },
-      }),
-      prisma.turn.findMany({ where: { matchId } }),
-    ]);
-    let local = selections.find((s: any) => s.userId === match.creatorId) || null;
-    let visitor = selections.find((s: any) => s.userId !== match.creatorId) || null;
-    if (!local || !visitor) {
-      if (match.creatorId) {
-        local =
-          selections.find((s: any) => s.userId === match.creatorId) ||
-          selections[0] ||
-          null;
-        visitor =
-          selections.find((s: any) => s.userId !== match.creatorId) ||
-          selections[1] ||
-          null;
-      } else {
-        local = selections[0] || null;
-        visitor = selections.length > 1 ? selections[1] : null;
-      }
-    }
-
-    const turnsCount = await prisma.turn.count({ where: { matchId } });
-    const half = turnsCount < 16 ? 1 : 2; // approximation
-    const turn = (turnsCount % 16) + 1; // approximation
-
-    const pickName = (sel: any) =>
-      sel?.teamRef?.name || sel?.teamRef?.roster || sel?.team || "";
-    const pickCoach = (sel: any) => sel?.user?.name || sel?.user?.email || "";
-    const pickElo = (sel: any) => sel?.user?.eloRating ?? 1000;
-    const acceptedUserIds = Array.from(
-      new Set(
-        (acceptTurns || [])
-          .map((t: any) =>
-            (t as any)?.payload?.type === "accept"
-              ? (t as any)?.payload?.userId
-              : null,
-          )
-          .filter(Boolean),
-      ),
-    );
-    const localAccepted = !!(local && acceptedUserIds.includes(local.userId));
-    const visitorAccepted = !!(
-      visitor && acceptedUserIds.includes(visitor.userId)
-    );
-
-    sendSuccess(res, {
-      id: match.id,
-      status: match.status,
-      createdAt: match.createdAt,
-      teams: {
-        local: { name: pickName(local), coach: pickCoach(local), eloRating: pickElo(local) },
-        visitor: { name: pickName(visitor), coach: pickCoach(visitor), eloRating: pickElo(visitor) },
-      },
-      score: { teamA: 0, teamB: 0 }, // TODO: remplacer par score réel quand disponible
-      half,
-      turn,
-      acceptances: { local: localAccepted, visitor: visitorAccepted },
-    });
-  } catch (e: unknown) {
-    serverLog.error(e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/:id/summary", authUser, handleGetMatchSummary);
+router.get("/details", handleGetMatchDetailsByTokenImpl);
+router.get("/:id/details", authUser, handleGetMatchDetailsImpl);
+router.get("/:id/teams", authUser, handleGetMatchTeamsImpl);
+router.get("/my-matches", authUser, handleListMyMatchesImpl);
+router.get("/:id/summary", authUser, handleGetMatchSummaryImpl);
 
 // S27.8.18 — `handleGetMatchState` extrait dans
 // `routes/match-state-handler.ts`. Re-export pour preserver l'API
